refactor: set map bounds and center dynamically
The map bounds and center were set based on markers and were made responsive to viewport change. This way the map would adapt to show all visible markers when visualized on mobile devices and when their orientation would change.
A README file was also included with a presentation of this project and credits to code snippets used to solve the current task.