refactor: Use KnockoutJS library to build the app
Rebuilded the app using the JavaScript library KnockoutJS, as requested in the project rubric.
The map is now very similar to the one that previously made only use of vanilla JavaScript, with the exception of the number of markers on selected locations, dropped from 5 to just 1.
The jQuery library was also used to apply the binding only when the HTML document is ready.