fix: Make the Google Maps API load before it is used
Fixed a bug that emerged every time the app was loaded for the first time.
As the Google Maps API hadn't finished loading yet, the same recurrent error 'google is not defined' showed up in the console. If the page was reloaded, the map displayed correctly in the page.
Defined an 'activateKO' callback function that instantiated the ViewModel and activated KO to make sure that the Google Maps API had finished loading before we used the script that depended on the Google Maps API.