feat: Add a distinct View Model for the list View
Added a distinct View Model for the View related to the list of places and a different custom binding.
That View is different and separated from the map View, yet it shares some similarities with it and it takes the same informations about places locations.
Added a link to a Stackoverflow answer I relied on upon building this feature in my code.