feat: Load map and update locations based on search
A major update was added with this commit. The KnockoutJS feature of Custom Binding Handlers was added to the app to take care of the complex functionality of updating the map locations based on the user search.
Custom Bindings would take care of loading the map as the app loads and of retrieving the search term the user would insert in the input form to look for specific locations.
The 'libraries' parameter was also included in the Maps API url to load the functionality of the Places service.