fix: Fix Wikipedia API problems
Fixed some major problems in displaying a Wikipedia link into each marker info window.
Dealt with cases where there was no response by the Wikipedia API by adding a custom message. Some changes to the code were made to make the query strings shorter by using the response array elements instead of composing the strings.
Fixed a bug that made the same link display into every info window when there was more than a single marker; this was due to a variable that was in an external scope and was assigned the last Google Places API result value.