fix: Fix marker color and infoWindow reopen
Fixed a problem with the assignment of the correct icon to markers. Before this fix, if an infoWindow was not closed and the user clicked another marker, both markers remained green and this behavior could even spread if other markers were clicked before closing the open infoWindow.
With this update, even if an infoWindow is not properly closed, only the selected marker gets a green icon.
Another problem involving the infoWindows was that after any of them was closed, it was not possible to reopen it before clicking to another one. This problem has been solved, as well.