feat: Make result list scrollable
Added code to make the result list with the locations scrollable if its height is bigger than the height of the div that hosts it.
The media queries added in the CSS file target some known devices and their official characteristics as of late 2016.
It is possible to add more devices or to adapt the code for changes in the future by modifying the height values.