fix: Prevent orientation changes by keyboard
Added a min-device-width to prevent that when the keyboard shows up in small devices it causes the map and list divs to be side by side.
In fact the keyboard causes the viewport to be resized and thus make the orientation appear as set to landscape in small devices. This triggers the shift of the position of the list div to be on the side of the map and makes it difficult to read.
Prevented this behavior for devices with width smaller than 500px.