refactor: Update custom binding and handle errors
Provided an error handling method to the Google Places API request in case the request is unsuccessful.
Refactored the custom binding method to use only the 'update' function, as the 'init' led to a duplicate function that ran as the app loaded the first time.
The former implementation also fired the error handling method described above for an unsuccessful request, because the 'update' function was also called the first time the binding was evaluated, leading to a request without a valid query.