refactor: Populate list view using Knockout template
Started to refactor the app to inprove the app architecture following the project requirements to avoid DOM manipulations.
Used the Knockout 'template' binding to populate the list View elements and simplified the map View using code from the Google Places SearchBox documentation instead of the custom binding as done before.