refactor: Provide array filter functionality
Provide a filter functionality that would filter (show or hide) the existing list of locations as well as the markers on the map.
After the API search has been successful, the user can perform a filter search on the returned results based on the matching query.
The filter search updates the view, but does not delete the saved results of the API search from the array where they are stored. It just populates the Knockout observable array with the matching results and empties it again if the filter string changes.
An empty filter query will result in showing again all the saved items returned by the API search.