refactor: Open infoWindow by clicking list item
Refactored the code to make the app open the info window on a marker if the user clicks on the equivalent list item.
Added a marker property that would take the value of the AJAX request response and store it to be called from the ViewModel as a result of a click on a list item.