fix: Highlight selected item and marker
Fixed a problem in highlighting only the selected list item and marker. After refactoring the code to show / hide markers instead of recreating them for performance reasons, there was a problem for the app to associate the selected marker to the appropriate item.
The problem became serious when the filter was used and a selected item should have losed its selection only if filtered out. Keeping it visible, instead, should have kept its selection.
Code was added to the filter function and to the item selection function to solve the issue.