fix: Set icon and item to normal if infowindow closes
Fixed a problem caused by closing an infowindow after filtering the list. If a selection was made after the list had been filtered, the marker kept a green icon from the selection without going back to normal.
Refactored the code to set only the selected marker icon to the green icon showing a selection. Deleted code that redrew all icons every time the list was filtered, to improve performance.