refactor: Delete unused code
Deleted commented code after a heavy refactoring on the two functions that take care of selecting the right list item and marker and populating the info window.

This work has been necessary because by avoiding to completely remove the markers and redraw them, the markers and places array were not anymore in sync with the locations array inside the ViewModel.
The places and markers array are now regenerated only when there is a query to the Google Places API by the user, while the locations array is regenerated every time the filter works. Because of this, the latter has a different number of elements than the first ones. So the selected marker and the selected place reference the same location, but take their data from arrays that are different both because of their size and composition.